Refactor: Move config and related modules to services
