--- app/config/config_defaults.json
@@ -1,168 +1,0 @@
-{
-    "general": {
-        "architecture": {
-            "mode": "threads"
-        },
-        "llm": {
-            "provider": "openai",
-            "third_party": {
-                "openai": {
-                    "options": {
-                        "model": "gpt-4o",
-                        "temperature": 0.7,
-                        "max_tokens": 512
-                    }
-                },
-                "huggingface_endpoint": {
-                    "options": {
-                        "endpoint_url": "",
-                        "model": "",
-                        "access_token": "",
-                        "temperature": 0.7,
-                        "max_tokens": 512
-                    }
-                }
-            },
-            "local": {
-                "huggingface_pipeline": {
-                    "options": {
-                        "model": "microsoft/DialoGPT-medium",
-                        "temperature": 0.7,
-                        "torch_dtype": "auto"
-                    }
-                },
-                "llama_cpp": {
-                    "options": {
-                        "model_path": "",
-                        "temperature": 1.0,
-                        "max_tokens": 512,
-                        "n_ctx": 2048,
-                        "n_gpu_layers": 0,
-                        "n_batch": 1000,
-                        "top_p": 0.95,
-                        "top_k": 64,
-                        "repeat_penalty": 1.0,
-                        "min_p": 0.0,
-                        "chat_format": "chatml-function-calling"
-                    }
-                }
-            }
-        },
-        "embeddings": {
-            "use_local": true
-        },
-        "speech_to_text": {
-            "language": "en",
-            "ambient_transcription": {
-                "enable": false,
-                "chunk_duration": 3.0,
-                "storage_path": "ambient_logs/",
-                "filter_short_transcriptions": true,
-                "min_transcription_length": 10
-            },
-            "audio_input": {
-                "device_index": null,
-                "sample_rate": 16000,
-                "channels": 1,
-                "chunk_size": 1024,
-                "format": "int16"
-            },
-            "wake_word": {
-                "enabled": true,
-                "backend": "oww",
-                "model_path": null,
-                "threshold": 0.5,
-                "inference_framework": "onnx"
-            },
-            "transcription": {
-                "vad_enabled": true,
-                "vad_threshold": 0.5,
-                "silence_duration_ms": 900,
-                "max_speech_duration_s": 30,
-                "realtime_model": {
-                    "enabled": true,
-                    "model_size": "tiny",
-                    "device": "cpu",
-                    "compute_type": "int8"
-                },
-                "accurate_model": {
-                    "enabled": true,
-                    "model_size": "base",
-                    "device": "cpu",
-                    "compute_type": "int8"
-                }
-            },
-            "coordinator": {
-                "session_timeout_s": 10.0,
-                "multi_turn_enabled": false,
-                "pause_tts_on_listen": true
-            }
-        },
-        "text_to_speech": {
-            "model_file_path": "/voice_models/en_US-lessac-medium.onnx",
-            "model_config_file_path": "/voice_models/en_US-lessac-medium.onnx.txt",
-            "model_sample_rate": 22050,
-            "piper_path": ""
-        },
-        "hardware_acceleration": {
-            "tts": false,
-            "stt": false,
-            "ocr_bg": false,
-            "ocr_curr": false,
-            "llm": false
-        }
-    },
-    "ui": {
-        "activate": false,
-        "dark_mode": false,
-        "debug": false
-    },
-    "plugins": {
-        "google": {
-            "credentials_file": "google_credentials.json"
-        },
-        "jira": {
-            "activate": false,
-            "api_token": "",
-            "username": "",
-            "instance_url": "https://jira.atlassian.net/"
-        },
-        "openrecall": {
-            "activate": false
-        },
-        "brave_search": {
-            "activate": false,
-            "api_key": ""
-        },
-        "github": {
-            "activate": false,
-            "app_id": "",
-            "app_private_key": "",
-            "repository": ""
-        },
-        "slack": {
-            "activate": false,
-            "user_token": ""
-        },
-        "gmail": {
-            "activate": false
-        },
-        "gcalendar": {
-            "activate": false
-        }
-    },
-    "mcp": {
-        "enabled": true,
-        "servers": {}
-    },
-    "messaging": {
-        "redis": {
-            "url": "redis://localhost:6379"
-        },
-        "priorities": {
-            "interactive": 10,
-            "system": 50,
-            "external": 80
-        }
-    }
-}
