+++ app/services/db/manager.py
@@ -10,8 +10,8 @@ from typing import Optional
 
 import aiosqlite
 
-from app.db.migration_manager import MigrationManager
-from app.db.models import Message
+from app.services.db.migration_manager import MigrationManager
+from app.services.db.models import Message
 from app.helpers.aurora_logger import log_error, log_info
