+++ app/services/tooling/tools/__init__.py
@@ -3,19 +3,19 @@
 Contains all tool implementations including core tools and plugin tools.
 """
 
-from app.tooling.tools.pomodoro_tool import (
+from app.services.tooling.tools.pomodoro_tool import (
     pomodoro_status_tool,
     start_pomodoro_tool,
     stop_pomodoro_tool,
 )
-from app.tooling.tools.resume_tts import resume_tts_tool
-from app.tooling.tools.scheduler_tool import (
+from app.services.tooling.tools.resume_tts import resume_tts_tool
+from app.services.tooling.tools.scheduler_tool import (
     cancel_scheduled_task_tool,
     list_scheduled_tasks_tool,
     schedule_task_tool,
 )
-from app.tooling.tools.stop_tts import stop_tts_tool
-from app.tooling.tools.upsert_memory import upsert_memory_tool
+from app.services.tooling.tools.stop_tts import stop_tts_tool
+from app.services.tooling.tools.upsert_memory import upsert_memory_tool
 
 __all__ = [
     # Core tools
